+++ changelog.md
@@ -10,6 +10,7 @@
 - Fix shader based particles breaking on servers (BluSunrize)
 - Fix various button tooltips sticking around after being clicked (BluSunrize)
 - Fix treated trapdoors erroneously having metal properties (BluSunrize)
+- Fix arms on posts dropping blocks when broken (BluSunrize)
 
 ##### Version 1.20.4-11.5.0-179
 - Add warning signs (BluSunrize)